feat(ldlog): add func `SetBriefStringLen` & `SetBriefArrayLen` & `SetBriefMapLen`
